backend/requirements.txt
Django==5.0.6
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
celery==5.4.0
redis==5.0.7
django-cors-headers==4.4.0
django-filter==24.2
Pillow==10.4.0
python-magic==0.4.27
